backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
supabase==2.10.0
PyPDF2==3.0.1
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==11.0.0
python-dotenv==1.0.1
openai==1.57.2
httpx==0.28.1
pydantic==2.10.3
pydantic-settings==2.6.1
reportlab==4.0.7
